Adding rows to tables
This is probably the most substantial change to the database yet. A
TON of bugs were encountered at this stage, and I think I've run up
hard against some of the assumptions I made at the beginning of this
process. Also, I've discovered that this stuff is _hard_. Like, crazy
hard.

Summary of changes:

- When a new table is created, we create a corresponding data block for
it.
- Each row in a data block is the same size
- The data follows a... really dumb structure where we always have n +
1 rows on the disk, where that +1 is a "sentinal" block which is
literally the exact same size as the rest of the rows but has a bit
flipped to indicate that it's the last row. Yes, this is dumb. Yes, I
plan on making a more robust system (probably involving btrees)

At this point I'm low-key kinda wanting to move onto a different piece
of the codebase, namely the query parser, so I'll be working on that
next. There's still a TON left to do ofc, and I'll add a todo list in
the next commit, but this is a... decent foundation. There's some
things that'll need to get refactored in the future, but I think I've
hidden their weirdness away from the rest of the codebase reasonably
well.