Query Parsing and Execution
Once again, we have a pretty substantial change to the database. One
of the things I'm pretty proud of is that the blockdisk and block
files were able to come out of this unscathed. That's good, because it
means those abstractions are working well.

There's definitely some weirdness to how things currently work. I
think I should have spent some time working on the Table concept
before starting on this task, but I also think that I now have a
really good idea of what Table needs to look like. Tradeoffs I guess

The approach for parsing is fairly simple. I don't really parse that
many constructs: create table has the most complex ast by far, but
it's also the one that I think will be the easiest to maintain in the
future. Inserting rows is... really weird. Semantic checking there
needs some serious work, and there are pretty bad gaps in it that
could lead to serious, unrecoverable bugs. So uhhhh... should probably
add an item in the todo list to fix them